requirements.txt
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
flexcache==0.3
flexparser==0.4
iniconfig==2.0.0
packaging==23.2
paramiko==3.5.0
Pint==0.24.4
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
pyfrc==2025.0.0
PyNaCl==1.5.0
pynetconsole==2.0.4
pynetworktables==2021.0.0
pyntcore==2025.2.1.2
pytest==8.3.4
pytest-reraise==2.1.2
robotpy==2025.2.1.3
robotpy-apriltag==2025.2.1
robotpy-cli==2024.0.0
robotpy-commands-v2==2025.1.1
robotpy-hal==2025.2.1.2
robotpy-halsim-gui==2025.2.1.2
robotpy-installer==2025.0.3
robotpy-navx==2025.0.1
robotpy-rev==2025.0.2
robotpy-wpilib-utilities==2025.0.0
robotpy-wpimath==2025.2.1.2
robotpy-wpinet==2025.2.1.2
robotpy-wpiutil==2025.2.1.2
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
wpilib==2025.2.1.2
